Bug 1588096 - combine lib into a single short file
